+++ src/account/types/balance.rs
@@ -5,7 +5,11 @@ use std::collections::HashMap;
 
 use iota_client::block::{
     dto::U256Dto,
-    output::{dto::TokenIdDto, AliasId, FoundryId, NftId, OutputId, TokenId},
+    output::{
+        dto::TokenIdDto,
+        feature::{dto::MetadataFeatureDto, MetadataFeature},
+        AliasId, FoundryId, NftId, OutputId, TokenId,
+    },
 };
 use primitive_types::U256;
 use serde::{Deserialize, Serialize};
@@ -156,6 +160,8 @@ pub struct NativeTokensBalance {
     /// Token id
     #[serde(rename = "tokenId")]
     pub token_id: TokenId,
+    /// Token foundry immutable metadata
+    pub metadata: Option<MetadataFeature>,
     /// Total amount
     pub total: U256,
     /// Balance that can currently be spent
@@ -166,6 +172,7 @@ impl Default for NativeTokensBalance {
     fn default() -> Self {
         Self {
             token_id: TokenId::null(),
+            metadata: None,
             total: U256::from(0u8),
             available: U256::from(0u8),
         }
@@ -178,6 +185,8 @@ pub struct NativeTokensBalanceDto {
     /// Token id
     #[serde(rename = "tokenId")]
     pub token_id: TokenIdDto,
+    /// Token foundry immutable metadata
+    pub metadata: Option<MetadataFeatureDto>,
     /// Total amount
     pub total: U256Dto,
     /// Balance that can currently be spent
@@ -188,6 +197,7 @@ impl From<&NativeTokensBalance> for NativeTokensBalanceDto {
     fn from(value: &NativeTokensBalance) -> Self {
         Self {
             token_id: TokenIdDto::from(&value.token_id),
+            metadata: None,
             total: U256Dto::from(&value.total),
             available: U256Dto::from(&value.available),
         }